Move `StrokeSet` into bellframe
