+++ bellframe/src/lib.rs
@@ -29,7 +29,7 @@ pub use parity::Parity;
 pub use place_not::{PlaceNot, PnBlock};
 pub use row::{same_stage_vec::SameStageVec, InvalidRowError, Row, RowBuf};
 pub use stage::{IncompatibleStages, Stage};
-pub use stroke::Stroke;
+pub use stroke::{Stroke, StrokeSet};
 pub use truth::Truth;
 pub use utils::run_len;
